% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_dependencies.R
\name{show_dependencies}
\alias{show_dependencies}
\title{Show tutorial dependencies}
\usage{
show_dependencies(package = "cpnr", tutorial)
}
\arguments{
\item{package}{Name of package, currently only cpnr}

\item{tutorial}{Name of tutorial}
}
\value{
\code{show_dependencies} will return a \code{character} string with all
package dependencies
}
\description{
Display the package dependencies of the specified tutorial in the cpnr package.
}
\examples{
show_dependencies(tutorial = "first_tutorial")

}
