% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_dependencies.R
\name{show_dependencies}
\alias{show_dependencies}
\title{Show tutorial dependencies}
\usage{
show_dependencies(tutorial, package = "cpnr")
}
\arguments{
\item{tutorial}{Name of tutorial}

\item{package}{Name of package, currently only cpnr}
}
\value{
\code{show_dependencies} will return a \code{character} string with all
package dependencies
}
\description{
Display the package dependencies of the specified tutorial in the cpnr package.
}
\examples{
show_dependencies(tutorial = "divdyn")

}
\seealso{
\code{show_description} to see the tutorial description and
\code{show_tutorials} to see all available tutorials.

Other show functions: 
\code{\link{show_description}()},
\code{\link{show_tutorials}()}
}
\concept{show functions}
