% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tutorials.R
\name{show_tutorials}
\alias{show_tutorials}
\title{Show available tutorials}
\usage{
show_tutorials(package = "cpnr")
}
\arguments{
\item{package}{Name of package, currently only cpnr}
}
\value{
\code{show_tutorials} will return a \code{data.frame} containing
"title" and "description".
}
\description{
Display which tutorials are currently available in the cpnr package.
}
\examples{
show_tutorials()

}
