% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_tutorial.R
\name{start_tutorial}
\alias{start_tutorial}
\title{Run a tutorial}
\usage{
start_tutorial(tutorial, package = "cpnr", shiny_args = NULL)
}
\arguments{
\item{tutorial}{Name of tutorial}

\item{package}{Name of package, currently only cpnr}

\item{shiny_args}{Additional arguments to forward to
\href{https://shiny.rstudio.com/reference/shiny/latest/runApp.html}{shiny::runApp}}
}
\description{
Run a specified tutorial in the cpnr package.
A web browser will be started automatically to view the document.
}
\details{
Note that, when finished, the running tutorial needs to be stopped
by clicking on the stop symbol in the R console.
}
\section{Warning}{

Before running \code{start_tutorial}, make sure to have all package
dependencies installed. You can see all dependencies using
\code{show_dependencies}
}

\examples{
\dontrun{
show_dependencies(tutorial = "divdyn")

start_tutorial(tutorial = "divdyn")
}


}
